Moved config definition files to a more relevant place
