+++ src/Drivers/SPI/SPI.hpp
@@ -12,7 +12,6 @@
 #pragma once
 
 #include "SPI_Engine.hpp"
-#include "structs/StructV1.hpp"
 #include "linux/spi/spidev.h"
 
 // ==============================================================================
