+++ src/Modules/EEPROM/address/address.cpp
@@ -1,0 +1,10 @@
+/**
+ * @file address.cpp
+ * @author l.heywang (leonard.heywang@gmail.com)
+ * @brief Source for the address module.
+ * @version 0.1
+ * @date 2024-10-17
+ *
+ * @copyright Copyright (c) 2024
+ *
+ */
