+++ src/Modules/EEPROM/address/address.hpp
@@ -1,0 +1,10 @@
+/**
+ * @file address.hpp
+ * @author l.heywang (leonard.heywang@gmail.com)
+ * @brief Define addressing tools for the EEPROM to use cell evently.
+ * @version 0.1
+ * @date 2024-10-17
+ *
+ * @copyright Copyright (c) 2024
+ *
+ */
