+++ src/Modules/EEPROM/structs/interfaces/struct_automated_interfaces.hpp
